Disambiguate cast from int to double or long double
svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=791622

+++ tools/polygonal/kpToolPolygonalBase.cpp
@@ -269,7 +269,7 @@ void kpToolPolygonalBase::applyModifiers ()
         // diagonal (dist from start maintained)
         else
         {
-            const double dist = sqrt (diffx * diffx + diffy * diffy);
+            const double dist = sqrt ((double)(diffx * diffx + diffy * diffy));
 
             #define sgn(a) ((a)<0?-1:1)
             // Round distances _before_ adding to any coordinate
